Updates to asDecimal to preserve decimal place significance in service list

+++ x-pack/plugins/apm/public/components/app/ServiceOverview/ServiceList/index.js
@@ -10,7 +10,7 @@ import styled from 'styled-components';
 import { RelativeLink } from '../../../../utils/url';
 import { fontSizes, truncate } from '../../../../style/variables';
 import TooltipOverlay from '../../../shared/TooltipOverlay';
-import { asMillisWithDefault, asInteger } from '../../../../utils/formatters';
+import { asMillisWithDefault, asDecimal } from '../../../../utils/formatters';
 import { ManagedTable } from '../../../shared/ManagedTable';
 
 // TODO: Consolidate these formatting helpers centrally
@@ -18,7 +18,7 @@ function formatNumber(value) {
   if (value === 0) {
     return '0';
   }
-  const formatted = asInteger(value);
+  const formatted = asDecimal(value);
   return formatted <= 0.1 ? '< 0.1' : formatted;
 }
